Some fixes.

+++ main-service/src/main/kotlin/tv/comnata/mainservice/entities/websocket/requests/RoomActionRequest.kt
@@ -1,6 +1,6 @@
 package tv.comnata.mainservice.entities.websocket.requests
 
 class RoomActionRequest {
-    val seekTime: Int? = null
+    val seekTime: Double? = null
     val type: String? = null
 }
